+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-*
+* Configurator `useMasks` calculation to comply with [ripe-sdk/#313](https://github.com/ripe-tech/ripe-sdk/pull/313)
 
 ## [0.15.0] - 2021-09-14